chore: enum 위치 변경, service impl 위치 변경
