+++ eatngo-core/src/main/kotlin/com/eatngo/store/domain/StoreSubscription.kt
@@ -69,13 +69,3 @@ class StoreSubscription(
         return deletedAt == null
     }
 }
-
-/**
- * 알림 유형
- */
-enum class NotificationType {
-    NEW_PRODUCT,         // 새 상품 등록 알림
-    DISCOUNT,            // 할인 알림
-    RESERVATION_UPDATE,  // 예약 상태 변경 알림
-    STORE_UPDATE         // 매장 정보 변경 알림
-}
